// src/app/modules/business-common/src/lib/components/journey-schedule/enums/translation-keys.enum.ts
export enum TranslationKeys {
  City_KeyNode = 'City.',
  City_BOG = 'City.BOG',
  City_GIG = 'City.GIG',
  City_MDE = 'City.MDE',
  City_PEI = 'City.PEI',
  Journey_Status_Cancelled = 'Journey.Status.Cancelled',
  Journey_Status_Confirmed = 'Journey.Status.Confirmed',
  Journey_Status_Delayed = 'Journey.Status.Delayed',
  Schedule_Connection_StopLabel = 'Schedule.Connection_StopLabel',
  Schedule_Connection_Title = 'Schedule.Connection_Title',
  Schedule_Direct = 'Schedule.Direct',
  Schedule_ExtraDay_AccessibleLabel = 'Schedule.ExtraDay_AccessibleLabel',
  Schedule_ExtraDay_Arrival_NDaysLater = 'Schedule.ExtraDay.Arrival_NDaysLater',
  Schedule_ExtraDay_Arrival_NextDay = 'Schedule.ExtraDay.Arrival_NextDay',
  Schedule_ExtraDay_Day_Label = 'Schedule.ExtraDay.Day_Label',
  Schedule_ExtraDay_Days_Label = 'Schedule.ExtraDay.Days_Label',
  Schedule_NewTimeOfArrival_Label = 'Schedule.NewTimeOfArrival_Label',
  Schedule_NewTimeOfDeparture_Label = 'Schedule.NewTimeOfDeparture_Label',
  Schedule_PreviousTimeOfArrival_Label = 'Schedule.PreviousTimeOfArrival_Label',
  Schedule_PreviousTimeOfDeparture_Label = 'Schedule.PreviousTimeOfDeparture_Label',
  Schedule_Stop = 'Schedule.Stop',
  Schedule_Stops = 'Schedule.Stops',
  Schedule_Total_TravelTime_Label = 'Schedule.Total_TravelTime_Label',
}